added services and fixed nav bar
