+++ pages/rockets/NavigationLinks.tsx
@@ -15,7 +15,7 @@ function NavigationLinks({ formattedData, selection }: IProps) {
           key={element.id}
           href="/rockets/[rocket]" as={`/rockets/${formattedData[index].id}`}
         >
-          <a className={`lg:py-3 xl:py-5 lg:px-10 xl:px-0 mr-10 xl:mr-0 xl:mb-4 text-2xl lg:text-base lg:rounded-lg lg:border-2 lg:border-transparent font-semibold text-white whitespace-nowrap ${
+          <a className={`lg:py-3 xl:py-5 w-full text-center	mr-10 xl:mr-0 xl:mb-4 text-2xl lg:text-base lg:rounded-lg lg:border-2 lg:border-transparent font-semibold text-white whitespace-nowrap ${
             selection === index
               ? "lg:bg-yellow-500 text-opacity-100 lg:bg-opacity-50"
               : "text-opacity-50 lg:text-opacity-100"
